Move empty menu to menus module
